+++ examples/twitter/common/components/Application.jsx
@@ -1,10 +1,25 @@
-import PropTypes from 'prop-types';
 import React from 'react';
-
+import PropTypes from 'prop-types';
+import Spinner from 'react-spinner';
+import BemHelper from 'react-bem-helper';
+import { connect } from 'react-redux';
+import { getIsPending, getIsLoading } from '@flowio/redux-fetch';
 import Header from './Header';
 
-const Application = ({ children }) => (
-  <div>
+if (process.browser) {
+  require('react-spinner/react-spinner.css'); // eslint-disable-line global-require
+  require('./Application.css'); // eslint-disable-line global-require
+}
+
+const classes = new BemHelper('Application');
+
+const Application = ({ children, loading }) => (
+  <div {...classes()}>
+    {loading && (
+      <div {...classes('spinner')}>
+        <Spinner />
+      </div>
+    )}
     <Header />
     {children}
   </div>
@@ -14,6 +29,11 @@ Application.displayName = 'Application';
 
 Application.propTypes = {
   children: PropTypes.node,
+  loading: PropTypes.bool,
 };
 
-export default Application;
+const mapStateToProps = state => ({
+  loading: getIsLoading(state) || getIsPending(state),
+});
+
+export default connect(mapStateToProps)(Application);