+++ examples/twitter/common/components/Timeline.jsx
@@ -1,12 +1,12 @@
-import PropTypes from 'prop-types';
 import React from 'react';
+import PropTypes from 'prop-types';
 import BemHelper from 'react-bem-helper';
 import map from 'lodash/map';
 import TimelineItem from './TimelineItem';
 import { timelineShape } from '../utilities/propTypes';
 
 if (process.browser) {
-  require('./Timeline.css');
+  require('./Timeline.css'); // eslint-disable-line global-require
 }
 
 const classes = new BemHelper('Timeline');