+++ src/ActionTypes.js
@@ -1,0 +1,7 @@
+const ActionTypes = {
+  FETCH_FAILURE: '@@fetch/FETCH_FAILURE',
+  FETCH_REQUEST: '@@fetch/FETCH_REQUEST',
+  FETCH_SUCCESS: '@@fetch/FETCH_SUCCESS',
+};
+
+export default ActionTypes;