+++ src/isError.js
@@ -1,0 +1,3 @@
+export default function isError(candidate) {
+  return candidate instanceof Error;
+}